+++ HISTORY.rst
@@ -2,6 +2,8 @@ Next version
 ------------
 
 - Expose information about top concepts.
+- Fix a bug with returning concepts and collections not on the first page
+  of data through the Range header. (#3)
 
 0.2.0 (2014-05-14)
 ------------------